# src/main/java/docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Channel connector APIs"
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:7777"
    }
  ],
  "paths": {
...
    "/accounts/validate/{primaryIdentifierName}/{primaryIdentifierVal}": {
      "post": {
        "operationId": "validation",
        "parameters": [
          {
            "in": "header",
            "name": "amsUrl",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "amsName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "accountHoldingInstitutionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Content-Type",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "primaryIdentifierName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "primaryIdentifierVal",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponseDTO"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "validation-api-controller"
        ]
      }
    },
...
    "/channel/transaction/{transactionId}/resolve": {
      "post": {
        "operationId": "transactionResolve",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "transaction-api-controller"
        ]
      }
    },
    "/channel/transactionRequest": {
      "post": {
        "operationId": "transaction",
        "parameters": [
          {
            "in": "header",
            "name": "Platform-TenantId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CorrelationID",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionChannelRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "example": {
                  "transactionId": "c3ce2b92-85e1-42b1-a71f-08f2ee44d42b"
                }
              }
            },
            "description": "Accepted: Transaction id generated"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "Invalid Headers": {
                    "description": "Invalid Headers",
                    "value": {
                      "errorCategory": "Validation",
                      "errorCode": "error.msg.header.validation.errors",
                      "errorDescription": "The headers are invalid",
                      "developerMessage": "The headers are invalid",
                      "defaultUserMessage": "The headers are invalid",
                      "errors": [
                        {
                          "errorCategory": "Validation",
                          "errorCode": "error.msg.schema.platform.tenant.id.cannot.be.null.or.empty",
                          "errorDescription": "Platform Tenant Id cannot be null or empty"
                        },
                        {
                          "errorCategory": "Validation",
                          "errorCode": "error.msg.schema.platform.tenant.id.length.is.invalid",
                          "errorDescription": "Platform Tenant Id length is invalid",
                          "errorParameters": [
                            {
                              "key": "Platform-TenantId",
                              "value": "20"
                            }
                          ]
                        }
                      ]
                    }
                  },
                  "Invalid Request Body": {
                    "description": "Invalid Request Body",
                    "value": {
                      "errorCategory": "Validation",
                      "errorCode": "error.msg.schema.validation.errors",
                      "errorDescription": "The request is invalid",
                      "developerMessage": "The request is invalid",
                      "defaultUserMessage": "The request is invalid",
                      "errors": [
                        {
                          "errorCategory": "Validation",
                          "errorCode": "error.msg.schema.payer.cannot.be.null.or.empty",
                          "errorDescription": "Payer cannot be null or empty"
                        },
                        {
                          "errorCategory": "Validation",
                          "errorCode": "error.msg.schema.payee.cannot.be.null.or.empty",
                          "errorDescription": "Payee cannot be null or empty"
                        },
                        {
                          "errorCategory": "Validation",
                          "errorCode": "error.msg.schema.amount.cannot.be.null.or.empty",
                          "errorDescription": "Amount cannot be null or empty"
                        }
                      ]
                    }
                  }
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "tags": [
          "transaction-api-controller"
        ]
      }
    },
    "/channel/transfer": {
      "post": {
        "operationId": "transfer",
        "parameters": [
          {
            "in": "header",
            "name": "Platform-TenantId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-BatchID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CorrelationID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Registering-Institution-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionChannelRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "example": {
                  "transactionId": "c3ce2b92-85e1-42b1-a71f-08f2ee44d42b"
                }
              }
            },
            "description": "Accepted: Transaction id generated"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "Invalid Headers": {
                    "description": "Invalid Headers",
                    "value": {
                      "errorCategory": "Validation",
                      "errorCode": "error.msg.header.validation.errors",
                      "errorDescription": "The headers are invalid",
                      "developerMessage": "The headers are invalid",
                      "defaultUserMessage": "The headers are invalid",
                      "errors": [
                        {
                          "errorCategory": "Validation",
                          "errorCode": "error.msg.schema.platform.tenant.id.cannot.be.null.or.empty",
                          "errorDescription": "Platform Tenant Id cannot be null or empty"
                        },
                        {
                          "errorCategory": "Validation",
                          "errorCode": "error.msg.schema.platform.tenant.id.length.is.invalid",
                          "errorDescription": "Platform Tenant Id length is invalid",
                          "errorParameters": [
                            {
                              "key": "Platform-TenantId",
                              "value": "20"
                            }
                          ]
                        }
                      ]
                    }
                  },
                  "Invalid Request Body": {
                    "description": "Invalid Request Body",
                    "value": {
                      "errorCategory": "Validation",
                      "errorCode": "error.msg.schema.validation.errors",
                      "errorDescription": "The request is invalid",
                      "developerMessage": "The request is invalid",
                      "defaultUserMessage": "The request is invalid",
                      "errors": [
                        {
                          "errorCategory": "Validation",
                          "errorCode": "error.msg.schema.payer.cannot.be.null.or.empty",
                          "errorDescription": "Payer cannot be null or empty"
                        },
                        {
                          "errorCategory": "Validation",
                          "errorCode": "error.msg.schema.payee.cannot.be.null.or.empty",
                          "errorDescription": "Payee cannot be null or empty"
                        },
                        {
                          "errorCategory": "Validation",
                          "errorCode": "error.msg.schema.amount.cannot.be.null.or.empty",
                          "errorDescription": "Amount cannot be null or empty"
                        }
                      ]
                    }
                  }
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "tags": [
          "transfer-api-controller"
        ]
      }
    },
    "/channel/transfer/{transactionId}": {
      "get": {
        "operationId": "transferId",
        "parameters": [
          {
            "in": "path",
            "name": "transactionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Platform-TenantId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionStatusResponseDTO"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "transfer-api-controller"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComplexName": {
        "required": [
          "partyIdType",
          "partyIdentifier"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "fspId": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "partyIdType": {
            "type": "string",
            "enum": [
              "MSISDN",
              "EMAIL",
              "PERSONAL_ID",
              "BUSINESS",
              "DEVICE",
              "ACCOUNT_ID",
              "IBAN",
              "ALIAS"
            ]
          },
          "partyIdentifier": {
            "type": "string"
          },
          "partySubIdOrType": {
            "type": "string"
          }
        }
      },
      "CustomData": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "object"
          }
        }
      },
...
      "Extension": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "ExtensionList": {
        "type": "object",
        "properties": {
          "extension": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Extension"
            }
          }
        }
      },
...
      "GeoCode": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "string"
          },
          "longitude": {
            "type": "string"
          }
        }
      },
...
      "MoneyData": {
        "required": [
          "amount",
          "currency"
        ],
        "type": "object",
        "properties": {
          "amount": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          }
        }
      },
      "Party": {
        "required": [
          "partyIdInfo"
        ],
        "type": "object",
        "properties": {
          "merchantClassificationCode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "partyIdInfo": {
            "$ref": "#/components/schemas/PartyIdInfo"
          },
          "personalInfo": {
            "$ref": "#/components/schemas/PersonalInfo"
          }
        }
      },
      "PartyIdInfo": {
        "required": [
          "partyIdType",
          "partyIdentifier"
        ],
        "type": "object",
        "properties": {
          "fspId": {
            "type": "string"
          },
          "partyIdType": {
            "type": "string",
            "enum": [
              "MSISDN",
              "EMAIL",
              "PERSONAL_ID",
              "BUSINESS",
              "DEVICE",
              "ACCOUNT_ID",
              "IBAN",
              "ALIAS"
            ]
          },
          "partyIdentifier": {
            "type": "string"
          },
          "partySubIdOrType": {
            "type": "string"
          }
        }
      },
...
      "PersonalInfo": {
        "type": "object",
        "properties": {
          "complexName": {
            "$ref": "#/components/schemas/ComplexName"
          },
          "dateOfBirth": {
            "type": "string"
          }
        }
      },
...
      "Refund": {
        "type": "object",
        "properties": {
          "originalTransactionId": {
            "type": "string"
          },
          "refundReason": {
            "type": "string"
          }
        }
      },
...
      "TransactionChannelRequestDTO": {
        "required": [
          "amount",
          "payer"
        ],
        "type": "object",
        "properties": {
          "amount": {
            "$ref": "#/components/schemas/MoneyData"
          },
          "amountType": {
            "type": "string",
            "enum": [
              "SEND",
              "RECEIVE"
            ]
          },
          "clientRefId": {
            "type": "string"
          },
          "customData": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomData"
            }
          },
          "expiration": {
            "type": "string"
          },
          "extensionList": {
            "$ref": "#/components/schemas/ExtensionList"
          },
          "geoCode": {
            "$ref": "#/components/schemas/GeoCode"
          },
          "note": {
            "type": "string"
          },
          "payee": {
            "$ref": "#/components/schemas/Party"
          },
          "payer": {
            "$ref": "#/components/schemas/Party"
          },
          "transactionType": {
            "$ref": "#/components/schemas/TransactionType"
          }
        }
      },
      "TransactionStatusResponseDTO": {
        "type": "object",
        "properties": {
          "clientRefId": {
            "type": "string"
          },
          "completedTimestamp": {
            "type": "string",
            "format": "date-time",
            "writeOnly": true
          },
          "completedTimestampDate": {
            "type": "string",
            "format": "date-time"
          },
          "transactionId": {
            "type": "string"
          },
          "transferId": {
            "type": "string"
          },
          "transferState": {
            "type": "string",
            "enum": [
              "RECEIVED",
              "RESERVED",
              "COMMITTED",
              "ABORTED"
            ]
          }
        }
      },
      "TransactionType": {
        "type": "object",
        "properties": {
          "balanceOfPayments": {
            "type": "string"
          },
          "initiator": {
            "type": "string",
            "enum": [
              "PAYER",
              "PAYEE"
            ]
          },
          "initiatorType": {
            "type": "string",
            "enum": [
              "CONSUMER",
              "AGENT",
              "BUSINESS",
              "DEVICE"
            ]
          },
          "refundInfo": {
            "$ref": "#/components/schemas/Refund"
          },
          "scenario": {
            "type": "string",
            "enum": [
              "DEPOSIT",
              "WITHDRAWAL",
              "TRANSFER",
              "PAYMENT",
              "REFUND",
              "MPESA"
            ]
          },
          "subScenario": {
            "type": "string"
          }
        }
      },
...
      "ValidationRequestDTO": {
        "type": "object",
        "properties": {
          "customData": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomData"
            }
          },
          "primaryIdentifier": {
            "$ref": "#/components/schemas/CustomData"
          },
          "secondaryIdentifier": {
            "$ref": "#/components/schemas/CustomData"
          }
        }
      },
      "ValidationResponseDTO": {
        "type": "object",
        "properties": {
          "accountHoldingInstitutionId": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "amsName": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "msisdn": {
            "type": "string"
          },
          "reconciled": {
            "type": "boolean"
          },
          "transactionId": {
            "type": "string"
          }
        }
      }
    }
  }
}